+++ packages/babel-preset-inglorious-script/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/babel-preset-inglorious-script
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/babel-plugin-inglorious-script@1.3.0
+
 ## 1.0.1
 
 ### Patch Changes